Permission structure is changed - Permission categories and actions created - 'scope' field added to roles - Lower scoped roles users can't access to higher scoped roles or users - Removed action links in list users page for logged user account, added profile link instead

+++ server/app/Http/Controllers/Api/v1/UserController.php
@@ -6,14 +6,20 @@ use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Hash;
 use App\Models\User;
 use App\Http\Resources\UserResource;
-use App\Http\Resources\UserCollection;
 use App\Http\Requests\StoreUserRequest;
 
 class UserController extends Controller
 {
-    public function index ():UserCollection
+    public function index ()
     {
-        return new UserCollection(User::all());
+        $users = User::whereHas('roles', function ($query) {
+            $loggedUser = auth()->user();
+            $loggedUserRoles = $loggedUser->roles->sortBy('scope');
+            $loggedUserScope = $loggedUserRoles[0]->scope;
+            return $query->where('scope', '>=', $loggedUserScope);
+        })->get();
+
+        return UserResource::collection($users);
     }
 
     public function store (StoreUserRequest $request):UserResource
@@ -27,6 +33,7 @@ class UserController extends Controller
         ]);
         $user->assignRole($request->roles);
         $user->givePermissionTo($request->permissions);
+
         return new UserResource($user);
     }
 
@@ -35,28 +42,50 @@ class UserController extends Controller
         return new UserResource($user);
     }
 
-    public function update (StoreUserRequest $request, User $user):UserResource
+    public function update (StoreUserRequest $request, User $user)
     {
-        $user->username = $request->username;
-        $user->email    = $request->email;
-        $user->name     = $request->name;
-        $user->surname  = $request->surname;
-        
-        if(!is_null($request->password)) {
-            $user->password = Hash::make($request->password);
+        $loggedUser = auth()->user();
+        $loggedUserRoles = $loggedUser->roles->sortBy('scope');
+        $loggedUserScope = $loggedUserRoles[0]->scope;
+        $userRoles = $user->roles->sortBy('scope');
+        $userScope = $userRoles[0]->scope;
+
+        if($loggedUserScope < $userScope) {
+            $user->username = $request->username;
+            $user->email    = $request->email;
+            $user->name     = $request->name;
+            $user->surname  = $request->surname;
+            
+            if(!is_null($request->password)) {
+                $user->password = Hash::make($request->password);
+            }
+            
+            $user->syncRoles($request->roles);
+            $user->syncPermissions($request->permissions);
+            $user->save();
+
+            return new UserResource($user);
+        } else {
+            return response(['message' => 'Forbidden'], 403);
         }
-        
-        $user->syncRoles($request->roles);
-        $user->syncPermissions($request->permissions);
-        $user->save();
-        return new UserResource($user);
     }
 
-    public function destroy (User $user):bool
+    public function destroy (User $user)
     {
-        $user->syncRoles();
-        $user->syncPermissions();
-        $user->delete();
-        return true;
+        $loggedUser = auth()->user();
+        $loggedUserRoles = $loggedUser->roles->sortBy('scope');
+        $loggedUserScope = $loggedUserRoles[0]->scope;
+        $userRoles = $user->roles->sortBy('scope');
+        $userScope = $userRoles[0]->scope;
+
+        if($loggedUserScope < $userScope) {
+            $user->syncRoles();
+            $user->syncPermissions();
+            $user->delete();
+
+            return true;
+        } else {
+            return response(['message' => 'Forbidden'], 403);
+        }
     }
 }
